# docs/swagger.yaml
definitions:
  models.Clip:
    properties:
      creator:
        $ref: '#/definitions/models.Creator'
      creator_uuid:
        type: string
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      game:
        $ref: '#/definitions/models.Game'
      game_uuid:
        type: string
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
      viewcount:
        type: integer
      vod:
        $ref: '#/definitions/models.Vod'
      vod_offset:
        type: integer
      vod_uuid:
        type: string
    type: object
  models.Creator:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
...
  models.Game:
    properties:
      box_art_url:
        type: string
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
  models.Vod:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      fps:
        type: number
      publish:
        type: boolean
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /vods/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a vod
        in: query
        name: uuid
        type: string
      - description: The title of a vod
        in: query
        name: title
        type: string
      - description: The duration of a vod
        in: query
        name: duration
        type: integer
      - description: The date of a vod
        in: query
        name: date
        type: string
      - description: The filename of a vod
        in: query
        name: filename
        type: string
      - description: The resolution of a vod
        in: query
        name: resolution
        type: string
      - description: The fps of a vod
        in: query
        name: fps
        type: integer
      - description: The size of a vod
        in: query
        name: size
        type: integer
      - description: 'Set order direction divided by comma. Possible ordering values:
          ''date'', ''duration'', ''size''. Possible directions: ''asc'', ''desc''.
          Example: ''date,desc'''
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vod'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all vods
      tags:
      - Vods
    post:
      consumes:
      - application/json
      parameters:
      - description: Vod obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Vod'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create vod
      tags:
      - Vods
  /vods/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete vod
      tags:
      - Vods
    get:
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vod'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get vod by uuid
      tags:
      - Vods
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      - description: Vod obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Vod'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Patch vod
      tags:
      - Vods
...
swagger: "2.0"